sort infer diffusion
