+++ examples/dit/diffusion/__init__.py
@@ -3,7 +3,7 @@
 #     ADM:   https://github.com/openai/guided-diffusion/blob/main/guided_diffusion
 #     IDDPM: https://github.com/openai/improved-diffusion/blob/main/improved_diffusion/gaussian_diffusion.py
 
-from . import gaussian_diffusion as gd
+from .diffusion_utils import LossType, ModelMeanType, ModelVarType, get_named_beta_schedule
 from .respace import SpacedDiffusion, space_timesteps
 
 
@@ -17,23 +17,23 @@ def create_diffusion(
     rescale_learned_sigmas=False,
     diffusion_steps=1000,
 ):
-    betas = gd.get_named_beta_schedule(noise_schedule, diffusion_steps)
+    betas = get_named_beta_schedule(noise_schedule, diffusion_steps)
     if use_kl:
-        loss_type = gd.LossType.RESCALED_KL
+        loss_type = LossType.RESCALED_KL
     elif rescale_learned_sigmas:
-        loss_type = gd.LossType.RESCALED_MSE
+        loss_type = LossType.RESCALED_MSE
     else:
-        loss_type = gd.LossType.MSE
+        loss_type = LossType.MSE
     if timestep_respacing is None or timestep_respacing == "":
         timestep_respacing = [diffusion_steps]
     return SpacedDiffusion(
         use_timesteps=space_timesteps(diffusion_steps, timestep_respacing),
         betas=betas,
-        model_mean_type=(gd.ModelMeanType.EPSILON if not predict_xstart else gd.ModelMeanType.START_X),
+        model_mean_type=(ModelMeanType.EPSILON if not predict_xstart else ModelMeanType.START_X),
         model_var_type=(
-            (gd.ModelVarType.FIXED_LARGE if not sigma_small else gd.ModelVarType.FIXED_SMALL)
+            (ModelVarType.FIXED_LARGE if not sigma_small else ModelVarType.FIXED_SMALL)
             if not learn_sigma
-            else gd.ModelVarType.LEARNED_RANGE
+            else ModelVarType.LEARNED_RANGE
         ),
         loss_type=loss_type,
     )
