+++ examples/dit/diffusion/gaussian_diffusion.py
@@ -4,8 +4,6 @@
 #     IDDPM: https://github.com/openai/improved-diffusion/blob/main/improved_diffusion/gaussian_diffusion.py
 
 
-import enum
-import math
 from functools import partial
 
 import numpy as np
@@ -15,131 +13,8 @@ from mindspore import Tensor
 from mindspore import dtype as mstype
 from mindspore import ops
 
-from .diffusion_utils import discretized_gaussian_log_likelihood, normal_kl
-
-
-def mean_flat(tensor):
-    """
-    Take the mean over all non-batch dimensions.
-    """
-    return tensor.mean(axis=list(range(1, len(tensor.shape))))
-
-
-class ModelMeanType(enum.Enum):
-    """
-    Which type of output the model predicts.
-    """
-
-    PREVIOUS_X = enum.auto()  # the model predicts x_{t-1}
-    START_X = enum.auto()  # the model predicts x_0
-    EPSILON = enum.auto()  # the model predicts epsilon
-
-
-class ModelVarType(enum.Enum):
-    """
-    What is used as the model's output variance.
-    The LEARNED_RANGE option has been added to allow the model to predict
-    values between FIXED_SMALL and FIXED_LARGE, making its job easier.
-    """
-
-    LEARNED = enum.auto()
-    FIXED_SMALL = enum.auto()
-    FIXED_LARGE = enum.auto()
-    LEARNED_RANGE = enum.auto()
-
-
-class LossType(enum.Enum):
-    MSE = enum.auto()  # use raw MSE loss (and KL when learning variances)
-    RESCALED_MSE = enum.auto()  # use raw MSE loss (with RESCALED_KL when learning variances)
-    KL = enum.auto()  # use the variational lower-bound
-    RESCALED_KL = enum.auto()  # like KL, but rescale to estimate the full VLB
-
-    def is_vb(self):
-        return self == LossType.KL or self == LossType.RESCALED_KL
-
-
-def _warmup_beta(beta_start, beta_end, num_diffusion_timesteps, warmup_frac):
-    betas = beta_end * np.ones(num_diffusion_timesteps, dtype=np.float64)
-    warmup_time = int(num_diffusion_timesteps * warmup_frac)
-    betas[:warmup_time] = np.linspace(beta_start, beta_end, warmup_time, dtype=np.float64)
-    return betas
-
-
-def get_beta_schedule(beta_schedule, *, beta_start, beta_end, num_diffusion_timesteps):
-    """
-    This is the deprecated API for creating beta schedules.
-    See get_named_beta_schedule() for the new library of schedules.
-    """
-    if beta_schedule == "quad":
-        betas = (
-            np.linspace(
-                beta_start**0.5,
-                beta_end**0.5,
-                num_diffusion_timesteps,
-                dtype=np.float64,
-            )
-            ** 2
-        )
-    elif beta_schedule == "linear":
-        betas = np.linspace(beta_start, beta_end, num_diffusion_timesteps, dtype=np.float64)
-    elif beta_schedule == "warmup10":
-        betas = _warmup_beta(beta_start, beta_end, num_diffusion_timesteps, 0.1)
-    elif beta_schedule == "warmup50":
-        betas = _warmup_beta(beta_start, beta_end, num_diffusion_timesteps, 0.5)
-    elif beta_schedule == "const":
-        betas = beta_end * np.ones(num_diffusion_timesteps, dtype=np.float64)
-    elif beta_schedule == "jsd":  # 1/T, 1/(T-1), 1/(T-2), ..., 1
-        betas = 1.0 / np.linspace(num_diffusion_timesteps, 1, num_diffusion_timesteps, dtype=np.float64)
-    else:
-        raise NotImplementedError(beta_schedule)
-    assert betas.shape == (num_diffusion_timesteps,)
-    return betas
-
-
-def get_named_beta_schedule(schedule_name, num_diffusion_timesteps):
-    """
-    Get a pre-defined beta schedule for the given name.
-    The beta schedule library consists of beta schedules which remain similar
-    in the limit of num_diffusion_timesteps.
-    Beta schedules may be added, but should not be removed or changed once
-    they are committed to maintain backwards compatibility.
-    """
-    if schedule_name == "linear":
-        # Linear schedule from Ho et al, extended to work for any number of
-        # diffusion steps.
-        scale = 1000 / num_diffusion_timesteps
-        return get_beta_schedule(
-            "linear",
-            beta_start=scale * 0.0001,
-            beta_end=scale * 0.02,
-            num_diffusion_timesteps=num_diffusion_timesteps,
-        )
-    elif schedule_name == "squaredcos_cap_v2":
-        return betas_for_alpha_bar(
-            num_diffusion_timesteps,
-            lambda t: math.cos((t + 0.008) / 1.008 * math.pi / 2) ** 2,
-        )
-    else:
-        raise NotImplementedError(f"unknown beta schedule: {schedule_name}")
-
-
-def betas_for_alpha_bar(num_diffusion_timesteps, alpha_bar, max_beta=0.999):
-    """
-    Create a beta schedule that discretizes the given alpha_t_bar function,
-    which defines the cumulative product of (1-beta) over time from t = [0,1].
-    :param num_diffusion_timesteps: the number of betas to produce.
-    :param alpha_bar: a lambda that takes an argument t from 0 to 1 and
-                      produces the cumulative product of (1-beta) up to that
-                      part of the diffusion process.
-    :param max_beta: the maximum beta to use; use values lower than 1 to
-                     prevent singularities.
-    """
-    betas = []
-    for i in range(num_diffusion_timesteps):
-        t1 = i / num_diffusion_timesteps
-        t2 = (i + 1) / num_diffusion_timesteps
-        betas.append(min(1 - alpha_bar(t2) / alpha_bar(t1), max_beta))
-    return np.array(betas)
+# from .diffusion_utils import discretized_gaussian_log_likelihood, normal_kl, mean_flat, _extract_into_tensor, ModelMeanType, ModelVarType, LossType
+from .diffusion_utils import ModelMeanType, ModelVarType, _extract_into_tensor
 
 
 class GaussianDiffusion:
@@ -506,349 +381,267 @@ class GaussianDiffusion:
             yield out
             img = out["sample"]
 
-    def ddim_sample(
-        self,
-        model,
-        x,
-        t,
-        clip_denoised=True,
-        denoised_fn=None,
-        cond_fn=None,
-        model_kwargs=None,
-        eta=0.0,
-    ):
-        """
-        Sample x_{t-1} from the model using DDIM.
-        Same usage as p_sample().
-        """
-        out = self.p_mean_variance(
-            model,
-            x,
-            t,
-            clip_denoised=clip_denoised,
-            denoised_fn=denoised_fn,
-            model_kwargs=model_kwargs,
-        )
-        if cond_fn is not None:
-            out = self.condition_score(cond_fn, out, x, t, model_kwargs=model_kwargs)
+    # def ddim_sample(
+    #     self,
+    #     model,
+    #     x,
+    #     t,
+    #     clip_denoised=True,
+    #     denoised_fn=None,
+    #     cond_fn=None,
+    #     model_kwargs=None,
+    #     eta=0.0,
+    # ):
+    #     """
+    #     Sample x_{t-1} from the model using DDIM.
+    #     Same usage as p_sample().
+    #     """
+    #     out = self.p_mean_variance(
+    #         model,
+    #         x,
+    #         t,
+    #         clip_denoised=clip_denoised,
+    #         denoised_fn=denoised_fn,
+    #         model_kwargs=model_kwargs,
+    #     )
+    #     if cond_fn is not None:
+    #         out = self.condition_score(cond_fn, out, x, t, model_kwargs=model_kwargs)
 
-        # Usually our model outputs epsilon, but we re-derive it
-        # in case we used x_start or x_prev prediction.
-        eps = self._predict_eps_from_xstart(x, t, out["pred_xstart"])
+    #     # Usually our model outputs epsilon, but we re-derive it
+    #     # in case we used x_start or x_prev prediction.
+    #     eps = self._predict_eps_from_xstart(x, t, out["pred_xstart"])
 
-        alpha_bar = _extract_into_tensor(self.alphas_cumprod, t, x.shape)
-        alpha_bar_prev = _extract_into_tensor(self.alphas_cumprod_prev, t, x.shape)
-        sigma = eta * ops.sqrt((1 - alpha_bar_prev) / (1 - alpha_bar)) * ops.sqrt(1 - alpha_bar / alpha_bar_prev)
-        # Equation 12.
-        noise = ops.randn_like(x)
-        mean_pred = out["pred_xstart"] * ops.sqrt(alpha_bar_prev) + ops.sqrt(1 - alpha_bar_prev - sigma**2) * eps
-        nonzero_mask = (t != 0).float().reshape(-1, *([1] * (len(x.shape) - 1)))  # no noise when t == 0
-        sample = mean_pred + nonzero_mask * sigma * noise
-        return {"sample": sample, "pred_xstart": out["pred_xstart"]}
+    #     alpha_bar = _extract_into_tensor(self.alphas_cumprod, t, x.shape)
+    #     alpha_bar_prev = _extract_into_tensor(self.alphas_cumprod_prev, t, x.shape)
+    #     sigma = eta * ops.sqrt((1 - alpha_bar_prev) / (1 - alpha_bar)) * ops.sqrt(1 - alpha_bar / alpha_bar_prev)
+    #     # Equation 12.
+    #     noise = ops.randn_like(x)
+    #     mean_pred = out["pred_xstart"] * ops.sqrt(alpha_bar_prev) + ops.sqrt(1 - alpha_bar_prev - sigma**2) * eps
+    #     nonzero_mask = (t != 0).float().reshape(-1, *([1] * (len(x.shape) - 1)))  # no noise when t == 0
+    #     sample = mean_pred + nonzero_mask * sigma * noise
+    #     return {"sample": sample, "pred_xstart": out["pred_xstart"]}
 
-    def ddim_reverse_sample(
-        self,
-        model,
-        x,
-        t,
-        clip_denoised=True,
-        denoised_fn=None,
-        cond_fn=None,
-        model_kwargs=None,
-        eta=0.0,
-    ):
-        """
-        Sample x_{t+1} from the model using DDIM reverse ODE.
-        """
-        assert eta == 0.0, "Reverse ODE only for deterministic path"
-        out = self.p_mean_variance(
-            model,
-            x,
-            t,
-            clip_denoised=clip_denoised,
-            denoised_fn=denoised_fn,
-            model_kwargs=model_kwargs,
-        )
-        if cond_fn is not None:
-            out = self.condition_score(cond_fn, out, x, t, model_kwargs=model_kwargs)
-        # Usually our model outputs epsilon, but we re-derive it
-        # in case we used x_start or x_prev prediction.
-        eps = (
-            _extract_into_tensor(self.sqrt_recip_alphas_cumprod, t, x.shape) * x - out["pred_xstart"]
-        ) / _extract_into_tensor(self.sqrt_recipm1_alphas_cumprod, t, x.shape)
-        alpha_bar_next = _extract_into_tensor(self.alphas_cumprod_next, t, x.shape)
+    # def ddim_reverse_sample(
+    #     self,
+    #     model,
+    #     x,
+    #     t,
+    #     clip_denoised=True,
+    #     denoised_fn=None,
+    #     cond_fn=None,
+    #     model_kwargs=None,
+    #     eta=0.0,
+    # ):
+    #     """
+    #     Sample x_{t+1} from the model using DDIM reverse ODE.
+    #     """
+    #     assert eta == 0.0, "Reverse ODE only for deterministic path"
+    #     out = self.p_mean_variance(
+    #         model,
+    #         x,
+    #         t,
+    #         clip_denoised=clip_denoised,
+    #         denoised_fn=denoised_fn,
+    #         model_kwargs=model_kwargs,
+    #     )
+    #     if cond_fn is not None:
+    #         out = self.condition_score(cond_fn, out, x, t, model_kwargs=model_kwargs)
+    #     # Usually our model outputs epsilon, but we re-derive it
+    #     # in case we used x_start or x_prev prediction.
+    #     eps = (
+    #         _extract_into_tensor(self.sqrt_recip_alphas_cumprod, t, x.shape) * x - out["pred_xstart"]
+    #     ) / _extract_into_tensor(self.sqrt_recipm1_alphas_cumprod, t, x.shape)
+    #     alpha_bar_next = _extract_into_tensor(self.alphas_cumprod_next, t, x.shape)
 
-        # Equation 12. reversed
-        mean_pred = out["pred_xstart"] * ops.sqrt(alpha_bar_next) + ops.sqrt(1 - alpha_bar_next) * eps
+    #     # Equation 12. reversed
+    #     mean_pred = out["pred_xstart"] * ops.sqrt(alpha_bar_next) + ops.sqrt(1 - alpha_bar_next) * eps
 
-        return {"sample": mean_pred, "pred_xstart": out["pred_xstart"]}
+    #     return {"sample": mean_pred, "pred_xstart": out["pred_xstart"]}
 
-    def ddim_sample_loop(
-        self,
-        model,
-        shape,
-        noise=None,
-        clip_denoised=True,
-        denoised_fn=None,
-        cond_fn=None,
-        model_kwargs=None,
-        progress=False,
-        eta=0.0,
-    ):
-        """
-        Generate samples from the model using DDIM.
-        Same usage as p_sample_loop().
-        """
-        final = None
-        for sample in self.ddim_sample_loop_progressive(
-            model,
-            shape,
-            noise=noise,
-            clip_denoised=clip_denoised,
-            denoised_fn=denoised_fn,
-            cond_fn=cond_fn,
-            model_kwargs=model_kwargs,
-            progress=progress,
-            eta=eta,
-        ):
-            final = sample
-        return final["sample"]
+    # def ddim_sample_loop(
+    #     self,
+    #     model,
+    #     shape,
+    #     noise=None,
+    #     clip_denoised=True,
+    #     denoised_fn=None,
+    #     cond_fn=None,
+    #     model_kwargs=None,
+    #     progress=False,
+    #     eta=0.0,
+    # ):
+    #     """
+    #     Generate samples from the model using DDIM.
+    #     Same usage as p_sample_loop().
+    #     """
+    #     final = None
+    #     for sample in self.ddim_sample_loop_progressive(
+    #         model,
+    #         shape,
+    #         noise=noise,
+    #         clip_denoised=clip_denoised,
+    #         denoised_fn=denoised_fn,
+    #         cond_fn=cond_fn,
+    #         model_kwargs=model_kwargs,
+    #         progress=progress,
+    #         eta=eta,
+    #     ):
+    #         final = sample
+    #     return final["sample"]
 
-    def ddim_sample_loop_progressive(
-        self,
-        model,
-        shape,
-        noise=None,
-        clip_denoised=True,
-        denoised_fn=None,
-        cond_fn=None,
-        model_kwargs=None,
-        progress=False,
-        eta=0.0,
-    ):
-        """
-        Use DDIM to sample from the model and yield intermediate samples from
-        each timestep of DDIM.
-        Same usage as p_sample_loop_progressive().
-        """
-        assert isinstance(shape, (tuple, list))
-        if noise is not None:
-            img = noise
-        else:
-            img = ops.randn(
-                *shape,
-            )
-        indices = list(range(self.num_timesteps))[::-1]
+    # def ddim_sample_loop_progressive(
+    #     self,
+    #     model,
+    #     shape,
+    #     noise=None,
+    #     clip_denoised=True,
+    #     denoised_fn=None,
+    #     cond_fn=None,
+    #     model_kwargs=None,
+    #     progress=False,
+    #     eta=0.0,
+    # ):
+    #     """
+    #     Use DDIM to sample from the model and yield intermediate samples from
+    #     each timestep of DDIM.
+    #     Same usage as p_sample_loop_progressive().
+    #     """
+    #     assert isinstance(shape, (tuple, list))
+    #     if noise is not None:
+    #         img = noise
+    #     else:
+    #         img = ops.randn(
+    #             *shape,
+    #         )
+    #     indices = list(range(self.num_timesteps))[::-1]
 
-        if progress:
-            # Lazy import so that we don't depend on tqdm.
-            from tqdm.auto import tqdm
+    #     if progress:
+    #         # Lazy import so that we don't depend on tqdm.
+    #         from tqdm.auto import tqdm
 
-            indices = tqdm(indices)
+    #         indices = tqdm(indices)
 
-        for i in indices:
-            t = ms.Tensor([i] * shape[0])
-            # with th.no_grad():
-            out = self.ddim_sample(
-                model,
-                img,
-                t,
-                clip_denoised=clip_denoised,
-                denoised_fn=denoised_fn,
-                cond_fn=cond_fn,
-                model_kwargs=model_kwargs,
-                eta=eta,
-            )
-            yield out
-            img = out["sample"]
+    #     for i in indices:
+    #         t = ms.Tensor([i] * shape[0])
+    #         # with th.no_grad():
+    #         out = self.ddim_sample(
+    #             model,
+    #             img,
+    #             t,
+    #             clip_denoised=clip_denoised,
+    #             denoised_fn=denoised_fn,
+    #             cond_fn=cond_fn,
+    #             model_kwargs=model_kwargs,
+    #             eta=eta,
+    #         )
+    #         yield out
+    #         img = out["sample"]
 
-    def _vb_terms_bpd(self, model, x_start, x_t, t, clip_denoised=True, model_kwargs=None):
-        """
-        Get a term for the variational lower-bound.
-        The resulting units are bits (rather than nats, as one might expect).
-        This allows for comparison to other papers.
-        :return: a dict with the following keys:
-                 - 'output': a shape [N] tensor of NLLs or KLs.
-                 - 'pred_xstart': the x_0 predictions.
-        """
-        true_mean, _, true_log_variance_clipped = self.q_posterior_mean_variance(x_start=x_start, x_t=x_t, t=t)
-        out = self.p_mean_variance(model, x_t, t, clip_denoised=clip_denoised, model_kwargs=model_kwargs)
-        kl = normal_kl(true_mean, true_log_variance_clipped, out["mean"], out["log_variance"])
-        kl = mean_flat(kl) / np.log(2.0)
+    # def _vb_terms_bpd(self, model, x_start, x_t, t, clip_denoised=True, model_kwargs=None):
+    #     """
+    #     Get a term for the variational lower-bound.
+    #     The resulting units are bits (rather than nats, as one might expect).
+    #     This allows for comparison to other papers.
+    #     :return: a dict with the following keys:
+    #              - 'output': a shape [N] tensor of NLLs or KLs.
+    #              - 'pred_xstart': the x_0 predictions.
+    #     """
+    #     true_mean, _, true_log_variance_clipped = self.q_posterior_mean_variance(x_start=x_start, x_t=x_t, t=t)
+    #     out = self.p_mean_variance(model, x_t, t, clip_denoised=clip_denoised, model_kwargs=model_kwargs)
+    #     kl = normal_kl(true_mean, true_log_variance_clipped, out["mean"], out["log_variance"])
+    #     kl = mean_flat(kl) / np.log(2.0)
 
-        decoder_nll = -discretized_gaussian_log_likelihood(
-            x_start, means=out["mean"], log_scales=0.5 * out["log_variance"]
-        )
-        assert decoder_nll.shape == x_start.shape
-        decoder_nll = mean_flat(decoder_nll) / Tensor(np.log(2.0))
-        decoder_nll = decoder_nll.to(kl.dtype)
+    #     decoder_nll = -discretized_gaussian_log_likelihood(
+    #         x_start, means=out["mean"], log_scales=0.5 * out["log_variance"]
+    #     )
+    #     assert decoder_nll.shape == x_start.shape
+    #     decoder_nll = mean_flat(decoder_nll) / Tensor(np.log(2.0))
+    #     decoder_nll = decoder_nll.to(kl.dtype)
 
-        # At the first timestep return the decoder NLL,
-        # otherwise return KL(q(x_{t-1}|x_t,x_0) || p(x_{t-1}|x_t))
-        output = ops.where((t == 0), decoder_nll, kl)
-        return {"output": output, "pred_xstart": out["pred_xstart"]}
+    #     # At the first timestep return the decoder NLL,
+    #     # otherwise return KL(q(x_{t-1}|x_t,x_0) || p(x_{t-1}|x_t))
+    #     output = ops.where((t == 0), decoder_nll, kl)
+    #     return {"output": output, "pred_xstart": out["pred_xstart"]}
 
-    def training_losses(self, model, x_start, t, model_kwargs=None, noise=None):
-        """
-        Compute training losses for a single timestep.
-        :param model: the model to evaluate loss on.
-        :param x_start: the [N x C x ...] tensor of inputs.
-        :param t: a batch of timestep indices.
-        :param model_kwargs: if not None, a dict of extra keyword arguments to
-            pass to the model. This can be used for conditioning.
-        :param noise: if specified, the specific Gaussian noise to try to remove.
-        :return: a dict with the key "loss" containing a tensor of shape [N].
-                 Some mean or variance settings may also have other keys.
-        """
-        if model_kwargs is None:
-            model_kwargs = {}
-        if noise is None:
-            noise = ops.randn_like(x_start)
-        x_t = self.q_sample(x_start, t, noise=noise)
-        video_input = False
-        if x_t.dim() == 4:
-            B, C = x_t.shape[:2]
-        elif x_t.dim() == 5:
-            B, F, C = x_t.shape[:3]
-            video_input = True
-        else:
-            raise ValueError(f"Incorrect input shape. Expect to get 4 or 5 dimensional inputs, but got {x_t.dim()}")
-        terms = {}
+    # def training_losses(self, model, x_start, t, model_kwargs=None, noise=None):
+    #     """
+    #     Compute training losses for a single timestep.
+    #     :param model: the model to evaluate loss on.
+    #     :param x_start: the [N x C x ...] tensor of inputs.
+    #     :param t: a batch of timestep indices.
+    #     :param model_kwargs: if not None, a dict of extra keyword arguments to
+    #         pass to the model. This can be used for conditioning.
+    #     :param noise: if specified, the specific Gaussian noise to try to remove.
+    #     :return: a dict with the key "loss" containing a tensor of shape [N].
+    #              Some mean or variance settings may also have other keys.
+    #     """
+    #     if model_kwargs is None:
+    #         model_kwargs = {}
+    #     if noise is None:
+    #         noise = ops.randn_like(x_start)
+    #     x_t = self.q_sample(x_start, t, noise=noise)
+    #     video_input = False
+    #     if x_t.dim() == 4:
+    #         B, C = x_t.shape[:2]
+    #     elif x_t.dim() == 5:
+    #         B, F, C = x_t.shape[:3]
+    #         video_input = True
+    #     else:
+    #         raise ValueError(f"Incorrect input shape. Expect to get 4 or 5 dimensional inputs, but got {x_t.dim()}")
+    #     terms = {}
 
-        if self.loss_type == LossType.KL or self.loss_type == LossType.RESCALED_KL:
-            terms["loss"] = self._vb_terms_bpd(
-                model=model,
-                x_start=x_start,
-                x_t=x_t,
-                t=t,
-                clip_denoised=False,
-                model_kwargs=model_kwargs,
-            )["output"]
-            if self.loss_type == LossType.RESCALED_KL:
-                terms["loss"] *= self.num_timesteps
-        elif self.loss_type == LossType.MSE or self.loss_type == LossType.RESCALED_MSE:
-            model_output = model(x_t, t, **model_kwargs)
+    #     if self.loss_type == LossType.KL or self.loss_type == LossType.RESCALED_KL:
+    #         terms["loss"] = self._vb_terms_bpd(
+    #             model=model,
+    #             x_start=x_start,
+    #             x_t=x_t,
+    #             t=t,
+    #             clip_denoised=False,
+    #             model_kwargs=model_kwargs,
+    #         )["output"]
+    #         if self.loss_type == LossType.RESCALED_KL:
+    #             terms["loss"] *= self.num_timesteps
+    #     elif self.loss_type == LossType.MSE or self.loss_type == LossType.RESCALED_MSE:
+    #         model_output = model(x_t, t, **model_kwargs)
 
-            if self.model_var_type in [
-                ModelVarType.LEARNED,
-                ModelVarType.LEARNED_RANGE,
-            ]:
-                if not video_input:
-                    assert model_output.shape == (B, C * 2, *x_t.shape[2:])
-                    model_output, model_var_values = ops.split(model_output, C, axis=1)
-                    frozen_out = ops.cat([model_output.copy(), model_var_values], axis=1)
-                else:
-                    assert model_output.shape == (B, F, C * 2, *x_t.shape[3:])
-                    model_output, model_var_values = ops.split(model_output, C, axis=2)
-                    frozen_out = ops.cat([model_output.copy(), model_var_values], axis=2)
-                # Learn the variance using the variational bound, but don't let
-                # it affect our mean prediction.
-                terms["vb"] = self._vb_terms_bpd(
-                    model=lambda *args, r=frozen_out: r,
-                    x_start=x_start,
-                    x_t=x_t,
-                    t=t,
-                    clip_denoised=False,
-                )["output"]
-                if self.loss_type == LossType.RESCALED_MSE:
-                    # Divide by 1000 for equivalence with initial implementation.
-                    # Without a factor of 1/1000, the VB term hurts the MSE term.
-                    terms["vb"] *= self.num_timesteps / 1000.0
+    #         if self.model_var_type in [
+    #             ModelVarType.LEARNED,
+    #             ModelVarType.LEARNED_RANGE,
+    #         ]:
+    #             if not video_input:
+    #                 assert model_output.shape == (B, C * 2, *x_t.shape[2:])
+    #                 model_output, model_var_values = ops.split(model_output, C, axis=1)
+    #                 frozen_out = ops.cat([model_output.copy(), model_var_values], axis=1)
+    #             else:
+    #                 assert model_output.shape == (B, F, C * 2, *x_t.shape[3:])
+    #                 model_output, model_var_values = ops.split(model_output, C, axis=2)
+    #                 frozen_out = ops.cat([model_output.copy(), model_var_values], axis=2)
+    #             # Learn the variance using the variational bound, but don't let
+    #             # it affect our mean prediction.
+    #             terms["vb"] = self._vb_terms_bpd(
+    #                 model=lambda *args, r=frozen_out: r,
+    #                 x_start=x_start,
+    #                 x_t=x_t,
+    #                 t=t,
+    #                 clip_denoised=False,
+    #             )["output"]
+    #             if self.loss_type == LossType.RESCALED_MSE:
+    #                 # Divide by 1000 for equivalence with initial implementation.
+    #                 # Without a factor of 1/1000, the VB term hurts the MSE term.
+    #                 terms["vb"] *= self.num_timesteps / 1000.0
 
-            target = {
-                ModelMeanType.PREVIOUS_X: self.q_posterior_mean_variance(x_start=x_start, x_t=x_t, t=t)[0],
-                ModelMeanType.START_X: x_start,
-                ModelMeanType.EPSILON: noise,
-            }[self.model_mean_type]
-            assert model_output.shape == target.shape == x_start.shape
-            terms["mse"] = mean_flat((target - model_output) ** 2)
-            if "vb" in terms:
-                terms["loss"] = terms["mse"] + terms["vb"]
-            else:
-                terms["loss"] = terms["mse"]
-        else:
-            raise NotImplementedError(self.loss_type)
+    #         target = {
+    #             ModelMeanType.PREVIOUS_X: self.q_posterior_mean_variance(x_start=x_start, x_t=x_t, t=t)[0],
+    #             ModelMeanType.START_X: x_start,
+    #             ModelMeanType.EPSILON: noise,
+    #         }[self.model_mean_type]
+    #         assert model_output.shape == target.shape == x_start.shape
+    #         terms["mse"] = mean_flat((target - model_output) ** 2)
+    #         if "vb" in terms:
+    #             terms["loss"] = terms["mse"] + terms["vb"]
+    #         else:
+    #             terms["loss"] = terms["mse"]
+    #     else:
+    #         raise NotImplementedError(self.loss_type)
 
-        return terms
-
-    def _prior_bpd(self, x_start):
-        """
-        Get the prior KL term for the variational lower-bound, measured in
-        bits-per-dim.
-        This term can't be optimized, as it only depends on the encoder.
-        :param x_start: the [N x C x ...] tensor of inputs.
-        :return: a batch of [N] KL values (in bits), one per batch element.
-        """
-        batch_size = x_start.shape[0]
-        t = ms.Tensor([self.num_timesteps - 1] * batch_size)
-        qt_mean, _, qt_log_variance = self.q_mean_variance(x_start, t)
-        kl_prior = normal_kl(mean1=qt_mean, logvar1=qt_log_variance, mean2=0.0, logvar2=0.0)
-        return mean_flat(kl_prior) / np.log(2.0)
-
-    def calc_bpd_loop(self, model, x_start, clip_denoised=True, model_kwargs=None):
-        """
-        Compute the entire variational lower-bound, measured in bits-per-dim,
-        as well as other related quantities.
-        :param model: the model to evaluate loss on.
-        :param x_start: the [N x C x ...] tensor of inputs.
-        :param clip_denoised: if True, clip denoised samples.
-        :param model_kwargs: if not None, a dict of extra keyword arguments to
-            pass to the model. This can be used for conditioning.
-        :return: a dict containing the following keys:
-                 - total_bpd: the total variational lower-bound, per batch element.
-                 - prior_bpd: the prior term in the lower-bound.
-                 - vb: an [N x T] tensor of terms in the lower-bound.
-                 - xstart_mse: an [N x T] tensor of x_0 MSEs for each timestep.
-                 - mse: an [N x T] tensor of epsilon MSEs for each timestep.
-        """
-        batch_size = x_start.shape[0]
-
-        vb = []
-        xstart_mse = []
-        mse = []
-        for t in list(range(self.num_timesteps))[::-1]:
-            t_batch = ms.Tensor([t] * batch_size)
-            noise = ops.randn_like(x_start)
-            x_t = self.q_sample(x_start=x_start, t=t_batch, noise=noise)
-            # Calculate VLB term at the current timestep
-            # with th.no_grad():
-            out = self._vb_terms_bpd(
-                model,
-                x_start=x_start,
-                x_t=x_t,
-                t=t_batch,
-                clip_denoised=clip_denoised,
-                model_kwargs=model_kwargs,
-            )
-            vb.append(out["output"])
-            xstart_mse.append(mean_flat((out["pred_xstart"] - x_start) ** 2))
-            eps = self._predict_eps_from_xstart(x_t, t_batch, out["pred_xstart"])
-            mse.append(mean_flat((eps - noise) ** 2))
-
-        vb = ops.stack(vb, axis=1)
-        xstart_mse = ops.stack(xstart_mse, axis=1)
-        mse = ops.stack(mse, axis=1)
-
-        prior_bpd = self._prior_bpd(x_start)
-        total_bpd = vb.sum(axis=1) + prior_bpd
-        return {
-            "total_bpd": total_bpd,
-            "prior_bpd": prior_bpd,
-            "vb": vb,
-            "xstart_mse": xstart_mse,
-            "mse": mse,
-        }
-
-
-def _extract_into_tensor(a, t, x_shape):
-    """
-    Extract values from a 1-D numpy array for a batch of indices.
-    :param a: the 1-D numpy array.
-    :param t: a tensor of indices into the array to extract.
-    :param x_shape: a larger shape of K dimensions with the batch
-                            dimension equal to the length of timesteps.
-    :return: a tensor of shape [batch_size, 1, ...] where the shape has K dims.
-    """
-    b = t.shape[0]
-    out = ops.GatherD()(a, -1, t)
-    return out.reshape(b, *((1,) * (len(x_shape) - 1)))
+    #     return terms
